+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2021-07-09
+
 ### Added
 
 - Add metrics for percentage of available IPs left in a subnet.
@@ -38,7 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Project copied from [aws-operator](https://github.com/giantswarm/aws-operator)
 
-[Unreleased]: https://github.com/giantswarm/aws-collector/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-collector/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/giantswarm/aws-collector/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/giantswarm/aws-collector/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/giantswarm/aws-collector/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/giantswarm/aws-collector/compare/v1.0.0...v1.1.0